Add utils with mk_str / mk_uni functions
