+++ utils.py
@@ -1,0 +1,29 @@
+import sys
+
+def mk_str(input):
+    if (sys.version_info > (3, 0)):  #py3
+        if isinstance(input, str):
+            return input
+        else:
+            input =  input.encode('utf-8')
+            return input
+    else:  #py2
+        if type(input) == unicode:
+            input =  input.encode('utf-8')
+            return input
+        else:
+            return input
+
+def mk_uni(input):
+    if (sys.version_info > (3, 0)):  #py3
+        if isinstance(input, str):
+            return input
+        else:
+            input =  input.decode('utf-8')
+            return input
+    else: #py2
+        if type(input) != unicode:
+            input =  input.decode('utf-8')
+            return input
+        else:
+            return input
